requirements.txt
asgiref==3.8.1
Django==5.0.8
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework-docs==0.1.7
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
jsonpickle==3.2.2
packaging==24.1
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.2
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
